requirements.txt
streamlit
pandas
numpy
Pillow
streamlit-plotly-events
plotly
opencv-python-headless
pip install pyTelegramBotAPI


